+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to Sourcegraph are documented in this file.
 - When the `DEPLOY_TYPE` environment variable is incorrectly specified, Sourcegraph now shuts down and logs an error message.
 - The `experimentalFeatures.canonicalURLRedirect` site config property now defaults to `enabled`. Set it to `disabled` to disable redirection to the `appURL` from other hosts.
 - Updating `maxReposToSearch` site config no longer requires a server restart to take effect.
+- The update check page no longer shows an error if you are using an insiders build. Insiders builds will now notify site administrators that updates are available 40 days after the release date of the installed build.
 
 ### Removed